feat: 채팅 메시지 읽음 상태 비동기 처리 구현

+++ backend/src/main/java/com/goodee/coreconnect/chat/event/MessageReadEvent.java
@@ -1,25 +1,29 @@
 package com.goodee.coreconnect.chat.event;
 
+/**
+ * 메시지 읽음 이벤트
+ * 비동기 처리를 위해 이벤트 방식으로 읽음 상태를 업데이트합니다.
+ */
 public class MessageReadEvent {
-    private final Long userId;
-    private final Long roomId;
-    private final Long messageId;
+    private final Integer userId;
+    private final Integer roomId;
+    private final Integer messageId;
 
-    public MessageReadEvent(Long userId, Long roomId, Long messageId) {
+    public MessageReadEvent(Integer userId, Integer roomId, Integer messageId) {
         this.userId = userId;
         this.roomId = roomId;
         this.messageId = messageId;
     }
 
-    public Long getUserId() {
+    public Integer getUserId() {
         return userId;
     }
 
-    public Long getRoomId() {
+    public Integer getRoomId() {
         return roomId;
     }
 
-    public Long getMessageId() {
+    public Integer getMessageId() {
         return messageId;
     }
 }
